Added "destroy" method
I've added a destroy method wich can be very helpful in some situations.

For example: if we are loading HTML content - which contains datepicker inputs -
dynamically, if some of this content needs to be removed, we should also destroy
any timepicker inside of it, in order to keep the DOM cleaner and removing
jQuery data and event listeners that are no longer needed.