Avoids construction of new instance before destroying
In previous commits regarding "destroy" method there was a special case which
was not being considered. If someones tries to call the destroy method on an
Element that hasn't been instantiated as a timepicker, the timepicker will be
instantiated first and then destroyed, consuming resources unnecessarily.

After this commit, calling the destroy method on an element that is not a
timepicker wont have any effect.