Update bootstrap-timepicker.js
I tried to use bootstrap-timepicker  with Masked Input Plugin but I had no success because the hour should be return always wit two digits, otherwise cannot be validet by mask input, so the value it is canceled. All the time when the user clicks the input, the value it is reset to empty and has a strange behaviour. I did not find any settings in timepicker to force to return always 2 digits for the hour. So I decided to make changes in the code. 
I this commit are the changes with a new option "hourTwoDigits" which is default set to false, to not affect the behaviour for those who already imlpemented timepicker.
 $('#timepicker5').timepicker({
        hourTwoDigits:true
});