Timepicker will now close its widget when tab out
If the timepicker widget is showing when the user
tabs out of the input control (either via
SHIFT+TAB or just TAB), the widget was still
showing on the screen.  This commit fixes it so
that the widget will close if it is open at the
time the user tabs out.